reshape structure
