+++ Cargo.toml
@@ -10,5 +10,8 @@ tokio-io = "0.1"
 serde_json = "1.0"
 pretty_env_logger = "0.2.0"
 log = "0.4"
-rustc-serialize = "0.3.24"
 libc = "0.2.43"
+serde = "1.0.80"
+serde_derive = "1.0.80"
+rustc-serialize = "0.3.24"
+tempfile = "3.0.4"
